requirements.txt
# Core dependencies
python-dotenv>=0.19.0
beautifulsoup4>=4.10.0
markdown>=3.3.6
pandas>=2.0.0  # Required for CSV file parsing
litellm>=0.8.1

# Error handling and retries
tenacity>=8.0.1

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# GUI
PyQt6>=6.4.0

# Web Interface
Flask>=2.3.0
Flask-WTF>=1.2.0
WTForms>=3.0.0
Werkzeug>=2.3.0

